Fix CORS handling for analytics endpoints
